% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CpGs}
\alias{CpGs}
\title{Data frame describing loci on the 27 and 450k arrays}
\usage{
data(CpGs)
}
\description{
Data frame describing loci on the 27 and 450k arrays.
}
\examples{
data(CpGs)
head(CpGs)
}
\keyword{datasets}
