% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLumiM-class.R
\name{estimateM}
\alias{estimateM}
\title{Estimate methylation M-value matrix}
\usage{
estimateM(methyLumiM, returnType = c("ExpressionSet", "matrix"), offset = 100)
}
\arguments{
\item{methyLumiM}{A [MethyLumiM-class] or `eSet` object, which includes
methylated and unmethylated probe intensities.}

\item{returnType}{Determines whether to return an `ExpressionSet`
(`MethyLumiM` in this case) or a matrix object.}

\item{offset}{Offset added to the methylated and unmethylated probe
intensities when estimating the M-value.}
}
\value{
A `MethyLumiM` or matrix object of methylation M-values.
}
\description{
Estimate the methylation M-value matrix from a `MethyLumiM` or `eSet`
object, which includes methylated and unmethylated probe intensities.
}
\details{
The M-value is the log2 ratio between the Illumina methylated and
unmethylated probe intensities. As variations of small intensities can cause
big changes in the ratio estimation, an offset is added to the methylated
and unmethylated probe intensities when estimating the M-value.

Please check the lumi package for more details of the `estimateM` function.
}
\references{
Du, P., Zhang, X, Huang, C.C., Jafari, N., Kibbe, W.A., Hou, L.,
  and Lin, S.M. (2010) 'Comparison of Beta-value and M-value methods for
  quantifying methylation levels by microarray analysis'.
}
\author{
Pan Du
}
\keyword{methods}
