% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylumiR.R
\name{extractBarcodeAndPosition}
\alias{extractBarcodeAndPosition}
\title{Extract the barcode and position information from a Sentrix ID}
\usage{
extractBarcodeAndPosition(sentrixids)
}
\arguments{
\item{sentrixids}{A character vector of sentrix IDs that look like
`1632405013_R001_C001`.}
}
\value{
A data.frame with five columns:
  \describe{
    \item{sentrix}{the sentrix ID}
    \item{row}{the sentrix row as it appears in the ID, e.g. `"R001"`}
    \item{column}{the sentrix column as it appears in the ID, e.g. `"C001"`}
    \item{rowNumber}{numeric, the sentrix row}
    \item{columnNumber}{numeric, the sentrix column}
  }
}
\description{
The sentrix IDs from an Illumina sentrix array contain positional
information that might be useful. This function simply extracts that
information from the ID itself.
}
\examples{
extractBarcodeAndPosition(c('12341234_R001_C001'))
}
\seealso{
[methylumiR()]
}
\author{
Sean Davis <seandavi@gmail.com>
}
\keyword{manip}
