% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFilter.R
\name{featureFilter}
\alias{featureFilter}
\alias{featureFilter,MethyLumiSet-method}
\alias{featureFilter,MethyLumiM-method}
\title{Annotation-based filtering of features (CpG sites) in a MethyLumiSet or
MethyLumiM object}
\usage{
featureFilter(eset, require.entrez=FALSE,
    require.GOBP=FALSE, require.GOCC=FALSE,
    require.GOMF=FALSE, exclude.ChrX=FALSE,
    require.closeToTSS=FALSE, range.DistToTSS=c(-500, 300),
    require.CpGisland=FALSE, ...)
}
\arguments{
\item{eset}{A `MethyLumiSet` or `MethyLumiM` object.}

\item{require.entrez}{If `TRUE`, filter out features without an Entrez Gene
ID annotation.}

\item{require.GOBP, require.GOCC, require.GOMF}{If `TRUE`, filter out features
whose target genes are not annotated to at least one GO term in the BP, CC
and MF ontology, respectively.}

\item{exclude.ChrX}{If `TRUE`, filter out features on chromosome X to avoid
gender effects.}

\item{require.closeToTSS}{If `TRUE`, filter out features that are not close
to a transcription start site (TSS). Features without an annotation of
distance to TSS will also be removed. Can only be used for the GoldenGate
platform.}

\item{range.DistToTSS}{Ignored if `require.closeToTSS` is `FALSE`. A numeric
vector of length 2 indicating the range of tolerable distance from the
transcription start site (TSS) in basepairs. Features whose distance to TSS
falls outside this range will be removed. The default is
\eqn{c(-500, 300)}, where \eqn{-500} is the distance to TSS from the left
and 300 the distance from the right.}

\item{require.CpGisland}{If `TRUE`, filter out features that are not in CpG
islands.}

\item{...}{Unused, but available for specializing methods.}
}
\value{
A list consisting of:
  \describe{
    \item{eset}{The filtered `MethyLumiSet` or `MethyLumiM` object.}
    \item{filter.log}{A list giving details of how many probe sets were
      removed for each annotation-based filtering step performed.}
  }
}
\description{
Features with insufficient annotation carry little value for subsequent data
analysis. The function `featureFilter` provides options for filtering
features (CpG sites) from a `MethyLumiSet` (or `MethyLumiM`) object based on
available annotation data.
}
\references{
R. Bourgon, R. Gentleman, W. Huber, \emph{Independent filtering
  increases power for detecting differentially expressed genes}, PNAS,
  vol. 107, no. 21, pp. 9546-9551.
}
\seealso{
[genefilter::nsFilter()]
}
\author{
Chao-Jen Wong \email{cwon2@fhcrc.org}
}
