% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylumiR.R
\name{getAssayDataNameSubstitutions}
\alias{getAssayDataNameSubstitutions}
\title{Return a data.frame of AssayData name substitutions}
\usage{
getAssayDataNameSubstitutions()
}
\value{
A data.frame with two columns, regex and replacement.
}
\description{
The Illumina methylation platforms use two distinct technologies, the
"goldengate" platform and the "infinium" platform. Each uses different file
formats as well as different assay chemistry. To make the downstream data
handling more straightforward and uniform between the two, a simple mapping
from the column names in the output files of the Illumina software converts
Red/Green or Cy5/Cy3 to unmethylated/methylated. This function returns that
mapping.
}
\details{
A file in the `extdata` directory called `substitutions.txt` contains two
columns. The function loads this file and uses the first column as a match
against column names in the data file (with the "sample part" removed). If
matched, the second column gives the replacement.
}
\examples{
getAssayDataNameSubstitutions()
}
\author{
Sean Davis <seandavi@gmail.com>
}
\keyword{IO}
