% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylumIDAT.R
\name{IDATtoMatrix}
\alias{IDATtoMatrix}
\title{process a single IDAT (just the mean intensities)}
\usage{
IDATtoMatrix(x, fileExts = list(Cy3 = "Grn", Cy5 = "Red"), idatPath = ".")
}
\arguments{
\item{x}{character(1)}

\item{fileExts}{named list}

\item{idatPath}{character(1)}
}
\description{
process a single IDAT (just the mean intensities)
}
