% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{methylumi-package}
\alias{methylumi-package}
\alias{methylumi}
\title{Handle Illumina methylation data}
\description{
This package contains a class structure for handling methylation data from
Illumina as well as utility functions for loading the data from files
generated by Illumina. Normalization that attempts to correct for dye bias
is also included.
}
\details{
Important data classes include: [MethyLumiSet-class] and
[MethyLumiQC-class], both of which are subsets of the [MethyLumi-class]
class, which is a subset of the `eSet` class.

A worked example of the use of the package can be found by typing
`openVignette()`.

A full listing of the available documentation can be obtained by typing
`help.start()` and selecting `methylumi` from the `Packages` link, or by
typing `library(help="methylumi")`.

If you use the `methylumIDAT` function or its out-of-band preprocessing
mechanisms in your work, a citation to the paper "Low-level processing of
Illumina Infinium DNA methylation beadarrays" by TJ Triche, DJ
Weisenberger, D Van Den Berg, KD Siegmund, and PW Laird, Nucleic Acids
Research, 2013, would be appreciated.

\tabular{ll}{ Package: \tab methylumi\cr Type: \tab Package\cr License:
\tab GPL\cr }
}
\seealso{
\code{\link[Biobase:Biobase-package]{Biobase}}
}
\author{
Sean Davis <seandavi@gmail.com>
}
\keyword{IO}
