% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripMethyLumiSet.R
\name{MethyLumi-strippers}
\alias{MethyLumi-strippers}
\alias{stripOOB}
\alias{stripBeadNs}
\alias{stripBeadSDs}
\alias{stripMethyLumiSet}
\title{Strip excessive probe-level data from MethyLumiSets}
\usage{
stripBeadNs(object)

stripBeadSDs(object)

stripOOB(object)

stripMethyLumiSet(object)
}
\arguments{
\item{object}{an object of class MethyLumi or a subclass.}
}
\value{
The object, with the corresponding assay data elements removed.
}
\description{
450k datasets with probe-level standard errors, out-of-band intensities and
bead numbers can become huge. These functions help to manage their growth in
memory, at least until preprocessing and QC is completed, whereupon the
summary data can be exported to a RangedData-based object of some sort for
integration.
}
\author{
Tim Triche, Jr. <tim.triche@gmail.com>
}
