\name{methylumiGenerics}
\alias{combine27k450k}
\alias{compare.chips}
\alias{controlData}
\alias{controlData<-}
\alias{cy3}
\alias{Cy3.N}
\alias{Cy3<-}
\alias{cy5}
\alias{Cy5.N}
\alias{Cy5<-}
\alias{designIItoMandU}
\alias{designItoMandU}
\alias{detection}
\alias{detection<-}
\alias{diagnostics}
\alias{getBarcodes}
\alias{getProtocolData}
\alias{IDATsToDFs}
\alias{IDATtoDF}
\alias{intensities.IB}
\alias{intensities.M}
\alias{intensities.OOB}
\alias{intensities.OOB.allelic}
\alias{intensities.U}
\alias{intensitiesByChannel}
\alias{lumiCSV}
\alias{lumIDAT}
\alias{mergeProbeDesigns}
\alias{methylated.N}
\alias{methylated.N<-}
\alias{methylumi.bgcorr}
\alias{methylumi.diagnostics}
\alias{methylumiCSV}
\alias{NChannelSetToMethyLumiSet}
\alias{negctls}
\alias{negctls.stderr}
\alias{negnorm}
\alias{normctls}
\alias{plotNAs}
\alias{plotNegOob}
\alias{probeNAs}
\alias{psummary}
\alias{pval.detect}
\alias{pval.detect<-}
\alias{qc.probe.plot}
\alias{readMethyLumIDAT}
\alias{sampleNAs}
\alias{subset27kas450k}
\alias{subset450kas27k}
\alias{subsetCommonProbes}
\alias{total.intensity}
\alias{unmethylated.N}
\alias{unmethylated.N<-}
\alias{dataType}
\alias{dataType<-}
\alias{betas,SummarizedExperiment-method}
\alias{methylated,MethylSet-method}
\alias{unmethylated,MethylSet-method}
\alias{betas,GenomicMethylSet-method}
\alias{betas,MethylSet-method}

\title{
  Generics defined in methylumi
}
\description{
  See the individual classes for details of methods.
}
\author{
  Sean Davis, Pan Du, and Tim Triche, Jr.
}
