% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylumiR.R
\name{methylumiR}
\alias{methylumiR}
\title{Load data from an Illumina methylation platform}
\usage{
methylumiR(filename, qcfile = NULL, sampleDescriptions = NULL, sep = NULL, ...)
}
\arguments{
\item{filename}{A filename of the excel-like file from BeadStudio.}

\item{qcfile}{A filename of the excel-like QC file from BeadStudio.}

\item{sampleDescriptions}{A data.frame that contains at least one column,
SampleID (case insensitive). This column MUST match the part of the column
headers before the `.Avg_Beta`, etc. Also, if a column called SampleLabel
(case insensitive) is present, it is used for sample labels, IF that
column contains unique identifiers.}

\item{sep}{Separator used in the BeadStudio (or GenomeStudio) output file.
If `NULL`, the function estimates it automatically.}

\item{...}{Passed into [utils::read.delim()].}
}
\value{
A [MethyLumiSet-class] object.
}
\description{
This function is useful for loading Illumina methylation data into a
[MethyLumiSet-class] object. Sample information can be supplied and will
then be incorporated into the resulting phenoData slot.
}
\details{
Autodetects the file format, reads the data, makes substitutions of
assayData element names for the various formats so that the method
accessors will work, maps `sampleDescriptions` appropriately, and returns a
`MethyLumiSet` object.
}
\examples{
## Read in sample information
samps <- read.table(system.file("extdata/samples.txt",
                                package = "methylumi"), sep = "\t", header = TRUE)
## Perform the actual data reading
## This is an example of reading data from an
## Sentrix Array format file (actually two files,
## one for data and one for QC probes)
mldat <- methylumiR(system.file('extdata/exampledata.samples.txt', package = 'methylumi'),
                    qcfile = system.file('extdata/exampledata.controls.txt', package = "methylumi"),
                    sampleDescriptions = samps)
mldat
}
\seealso{
[MethyLumiSet-class], [MethyLumiQC-class]
}
\author{
Sean Davis <seandavi@gmail.com>
}
\keyword{IO}
