% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mldat}
\alias{mldat}
\title{Example SAM format Illumina methylation dataset}
\usage{
data(mldat)
}
\description{
This is an example [MethyLumiSet-class] object.
}
\examples{
data(mldat)
}
\keyword{datasets}
