% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLumiSet-class.R
\name{plotSampleIntensities}
\alias{plotSampleIntensities}
\title{Plot the sample intensities}
\usage{
plotSampleIntensities(x,beta.cuts,s)
}
\arguments{
\item{x}{An object of class MethyLumi or a subclass.}

\item{beta.cuts}{Cutoffs for low and high beta values.}

\item{s}{Sample number to plot.}
}
\description{
The Illumina methylation platforms all show a significant dye bias. The
`plotSampleIntensities` method shows the density plots for the two channels,
allowing direct visualization of the effect.
}
\examples{
data(mldat)
plotSampleIntensities(mldat, s = 1)
}
\keyword{hplot}
