% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyLumiSet-class.R
\name{qcplot}
\alias{qcplot}
\alias{controlTypes}
\title{Methods for dealing with control data for Illumina methylation data}
\usage{
qcplot(object,controltype,...)
controlTypes(object,...)
}
\arguments{
\item{object}{An object of class [MethyLumiSet-class] or
[MethyLumiQC-class].}

\item{controltype}{A single character value representing the bead type to
plot from the quality control data. The available types are accessible via
the `controlTypes` method.}

\item{...}{Passed to the plot function.}
}
\description{
The `qcplot` function simply generates a plot of the control probe
information for a given `controlType`.
}
\details{
The descriptions of the various control types can be obtained from the
Illumina methylation user's guides.
}
\examples{
data(mldat)
controlTypes(mldat)
qcplot(mldat, controlTypes(mldat)[3])
}
\seealso{
[MethyLumiSet-class], [MethyLumiQC-class]
}
\author{
Sean Davis <seandavi@gmail.com>
}
\keyword{hplot}
