% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tcgaPipeline}
\alias{tcgaPipeline}
\title{Total convenience function for processing IDATs like tcga}
\usage{
tcgaPipeline(IDATs)
}
\arguments{
\item{IDATs}{character() of idat files}
}
\description{
Total convenience function for processing IDATs like tcga
}
