% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFilter.R
\name{varFilter}
\alias{varFilter}
\alias{varFilter,MethyLumiSet-method}
\alias{varFilter,MethyLumiM-method}
\title{Variation-based filtering of features (CpG sites) in a MethyLumiSet or
MethyLumiM object}
\usage{
varFilter(eset, var.func=IQR, var.cutoff=0.5, filterByQuantile=TRUE, ...)
}
\arguments{
\item{eset}{A `MethyLumiSet` or `MethyLumiM` object.}

\item{var.func}{The function used as the per-feature filtering statistic.}

\item{var.cutoff}{A numeric value indicating the cutoff value for variation.
If `filterByQuantile` is `TRUE`, features whose value of `var.func` is less
than the `var.cutoff`-quantile of all `var.func` values will be removed. If
`FALSE`, features whose values are less than `var.cutoff` will be removed.}

\item{filterByQuantile}{A logical indicating whether `var.cutoff` is to be
interpreted as a quantile of all `var.func` values (the default), or as an
absolute value.}

\item{...}{Unused, but available for specializing methods.}
}
\value{
A list consisting of:
  \describe{
    \item{eset}{The filtered `MethyLumiSet` or `MethyLumiM` object.}
    \item{filter.log}{Shows how many low-variance features were removed.}
  }
}
\description{
The function `varFilter` removes features exhibiting little variation across
samples. Such non-specific filtering can be advantageous for downstream data
analysis.
}
\details{
This function is a counterpart of the functions `nsFilter` and `varFilter`
available from the `genefilter` package. See R. Bourgon et al. (2010) and
[genefilter::nsFilter()] for detail.

It is proven that non-specific filtering, for which the criterion does not
depend on sample class, can increase the number of discoveries.
Inappropriate choice of test statistics, however, might have an adverse
effect. `limma`'s moderated \eqn{t}-statistic, for example, is based on an
empirical Bayes approach which models the conjugate prior of gene-level
variance with an inverse of \eqn{\chi^2} distribution scaled by observed
global variance. As the variance-based filtering removes the set of genes
with low variance, the scaled inverse \eqn{\chi^2} no longer provides a good
fit to the data passing the filter, causing the `limma` algorithm to produce
a posterior degree-of-freedom of infinity (Bourgon 2010). This leads to two
consequences: (i) the gene-level variance estimate will be ignored, and (ii)
the \eqn{p}-value will be overly optimistic (Bourgon 2010).
}
\examples{
  data(mldat)
  ## keep top 75 percent
  filt <- varFilter(mldat, var.cutoff=0.25)
  filt$filter.log
  dim(filt$eset)
}
\references{
R. Bourgon, R. Gentleman, W. Huber, \emph{Independent filtering
  increases power for detecting differentially expressed genes}, PNAS,
  vol. 107, no. 21, pp. 9546-9551, 2010.
}
\seealso{
[genefilter::nsFilter()]
}
\author{
Chao-Jen Wong \email{cwon2@fhcrc.org}
}
