% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylumIDAT.R
\name{IDATsToMatrices}
\alias{IDATsToMatrices}
\title{convert multiple idats to matrices}
\usage{
IDATsToMatrices(
  barcodes,
  fileExts = list(Cy3 = "Grn", Cy5 = "Red"),
  parallel = F,
  idatPath = "."
)
}
\arguments{
\item{barcodes}{character()}

\item{fileExts}{character()}

\item{parallel}{logical(1)}

\item{idatPath}{character(1)}
}
\description{
convert multiple idats to matrices
}
