\name{methylumIDAT}
\alias{methylumBAB}
\alias{methylumBLD}
\alias{methylumCSV}
\alias{lumIDAT}
\title{
  Raw data import for Illumina Infinium 27k and 450k methylation BeadArrays
}
\description{
  methylumIDAT() takes a list of 27k or 450k BeadArray barcodes, and returns 
  a MethyLumiSet containing the data.  lumIDAT() does the same except that it 
  creates a MethyLumiM object instead (this can be preferable for GEO submission
  and taking advantage of packages like GeneAnswers, though the conversion from
  one representation to the other is trivial now that the bugs are ironed out).

  The experimental methylumiBAB() and methylumiBLD() functions are hooks to use
  bead-level data (compressed or uncompressed via 'BeadDataPackR' or 'beadarray'
  respectively), and the original methylumiCSV() function (whence this entire 
  package arose) works on CSV files generated by certain configurations of iScan
}
\keyword{ package }
\seealso{
  \code{\link[methylumi:methylumi-package]{methylumi}}
  \code{\link[methylumi:MethyLumiSet-class]{MethyLumiSet}}
  \code{\link[methylumi:MethyLumiM-class]{MethyLumiM}}
}
\examples{

## AML <- methylumIDAT(AML.barcodes, parallel=TRUE)
## 
##   Processing IDAT files from 216 arrays... 
##   reading protocolData for 5318317007_A 
##   reading protocolData for 5318317007_B 
##   ...
##   reading protocolData for 5380348029_L 
##
## AML.lumi <- as(AML, 'MethyLumiM')
##
## head(protocolData(AML.idats)@data)
##                           ScanDate            DecodeDate
##  5318317007_A 5/25/2010 5:30:34 AM 3/24/2010 11:01:32 PM
##  5318317007_B 5/25/2010 5:35:40 AM 3/24/2010 11:01:56 PM
##  5318317007_C 5/25/2010 5:41:49 AM 3/24/2010 11:01:59 PM
##  5318317007_D 5/25/2010 5:47:20 AM 3/24/2010 11:02:04 PM
##  5318317007_E 5/25/2010 5:53:39 AM 3/24/2010 11:02:01 PM
##  5318317007_F 5/25/2010 5:59:18 AM 3/24/2010 11:03:11 PM
##
## summary(sampleNAs(AML)) # dropouts at pval > 0.05
##     Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
##        0       1       4     315      26   15080 
##
## produceMethylationGEOSubmissionFile(AML.lumi)

}
